Leaveo only client-relevant parts in lib/forwarders
